extract webhook debugging routes
