+++ lib/routes/products.rb
